requirements.txt
python-dotenv==1.0.0
pdf2image==1.16.3
requests==2.31.0
urllib3==2.0.4
